refactor(auth-ui): move constants to auth-core lib
