+++ libs/auth/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/constants';
+export * from './lib/misc';
 export type { LoginPayload } from './lib/schemas/login-payload-schema';
 export { LoginPayloadSchema } from './lib/schemas/login-payload-schema';
 export type { LoginResponse } from './lib/schemas/login-response-schema';
