+++ libs/auth/ui/src/index.ts
@@ -1,4 +1,3 @@
-export { credentialInputs, loginInputs, nameInputs } from './lib/constants';
 export { LoginForm } from './lib/login-form';
 export { LoginFormSkeleton } from './lib/login-form-skeleton';
 export { SignupForm } from './lib/signup-form';
