+++ libs/auth/ui/src/lib/login-form.tsx
@@ -3,12 +3,11 @@ import type { Control } from 'react-hook-form';
 
 import { Stack, Title } from '@mantine/core';
 
-import type { LoginPayload } from '@lihim/auth/core';
+import { loginInputs, LoginPayload } from '@lihim/auth/core';
 import { NAME_PASSWORD, TESTID_LOGIN_LOADING } from '@lihim/auth/core';
 import type { TVoidFn } from '@lihim/shared/core';
 import { ValidatedInput } from '@lihim/shared/ui';
 
-import { loginInputs } from './constants';
 import { FormFooter } from './form-footer';
 import { FormWrapper } from './form-wrapper';
